fix: Resolve final unnecessary_cast lint in distribution.rs

+++ crates/pufferlib/src/policy/distribution.rs
@@ -169,7 +169,7 @@ impl Distribution {
                 let probs = logits.softmax(-1, Kind::Float);
                 let log_probs = logits.log_softmax(-1, Kind::Float);
                 let entropy = -(probs * log_probs).sum_dim_intlist(
-                    Some(&[-1 as i64][..]),
+                    Some(&[-1_i64][..]),
                     false,
                     Kind::Float,
                 );
